+++ pkg/morph/client/container/delete.go
@@ -4,12 +4,14 @@ import (
 	"fmt"
 
 	"github.com/nspcc-dev/neofs-node/pkg/morph/client"
+	fschaincontracts "github.com/nspcc-dev/neofs-node/pkg/morph/contracts"
 )
 
 // DeletePrm groups parameters of Delete client operation.
 type DeletePrm struct {
 	cnr       []byte
 	signature []byte
+	key       []byte
 	token     []byte
 
 	client.InvokePrmOptional
@@ -23,6 +25,11 @@ func (d *DeletePrm) SetCID(cid []byte) {
 // SetSignature sets signature.
 func (d *DeletePrm) SetSignature(signature []byte) {
 	d.signature = signature
+}
+
+// SetKey sets public key.
+func (d *DeletePrm) SetKey(key []byte) {
+	d.key = key
 }
 
 // SetToken sets session token.
@@ -41,8 +48,8 @@ func (c *Client) Delete(p DeletePrm) error {
 	}
 
 	prm := client.InvokePrm{}
-	prm.SetMethod(deleteMethod)
-	prm.SetArgs(p.cnr, p.signature, p.token)
+	prm.SetMethod(fschaincontracts.RemoveContainerMethod)
+	prm.SetArgs(p.cnr, p.signature, p.key, p.token)
 	prm.InvokePrmOptional = p.InvokePrmOptional
 
 	// no magic bugs with notary requests anymore, this operation should
@@ -52,7 +59,15 @@ func (c *Client) Delete(p DeletePrm) error {
 
 	err := c.client.Invoke(prm)
 	if err != nil {
-		return fmt.Errorf("could not invoke method (%s): %w", deleteMethod, err)
+		if isMethodNotFoundError(err, fschaincontracts.RemoveContainerMethod) {
+			prm.SetMethod(deleteMethod)
+			prm.SetArgs(p.cnr, p.signature, p.token)
+			if err = c.client.Invoke(prm); err != nil {
+				return fmt.Errorf("could not invoke method (%s): %w", deleteMethod, err)
+			}
+			return nil
+		}
+		return fmt.Errorf("could not invoke method (%s): %w", fschaincontracts.RemoveContainerMethod, err)
 	}
 	return nil
 }